move PositiveF64 to its own module
Code move only. Retain the public constructor. We will eliminate its use
over the following commits.

Also, publicly expose the type in src/lib.rs, since we are going to start
enforcing invariants on it and making it generally useful.
