+++ src/primitives/positive_f64.rs
@@ -1,0 +1,29 @@
+// SPDX-License-Identifier: CC0-1.0
+
+//! Positive floats ("branch probabilities" for policies)
+
+use core::{cmp, hash};
+
+/// Ordered f64 for comparison.
+#[derive(Copy, Clone, PartialEq, Debug)]
+pub struct PositiveF64(pub f64);
+
+impl Eq for PositiveF64 {}
+
+// We could derive PartialOrd, but we can't derive Ord, and clippy wants us
+// to derive both or neither. Better to be explicit.
+impl PartialOrd for PositiveF64 {
+    fn partial_cmp(&self, other: &Self) -> Option<cmp::Ordering> { Some(self.cmp(other)) }
+}
+
+impl Ord for PositiveF64 {
+    fn cmp(&self, other: &Self) -> cmp::Ordering {
+        // will panic if given NaN
+        self.0.partial_cmp(&other.0).unwrap()
+    }
+}
+
+/// Hash required for using OrdF64 as key for hashmap
+impl hash::Hash for PositiveF64 {
+    fn hash<H: hash::Hasher>(&self, state: &mut H) { self.0.to_bits().hash(state); }
+}
